Add torch.no_grad to reduce memory consumption
Previously, calling model(batch) to get predictions on a batch
of images would use so much of my computers RAM that it would
slow to a crawl.

Using torch.no_grad() to get predictions removes this memory
problem entirely. I'm unsure why PyTorch does anything with
gradient descent when getting predictions on an image, but at
any rate, this solves the problem for now.