Organize and simplify prediction pipeline
Moved several helper functions to a utils.py file to
be imported into model.py.

Changed the implementation to choose a single input image
from the images dataset as the batch, runs that image
through the detection model, and prints the metrics
associated with the results.

Things to add/change are many, but first I need to implement
a function that displays/saves the resulting bounding boxes.